crud/requirements.txt
Flask==3.1.0
mysql-connector-python

